requirements.txt
allure-pytest==2.13.2
flake8==7.0.0
flake8-isort==6.1.1
isort==5.13.2
pytest==7.4.4
pyyaml==6.0.1
requests==2.31.0
scp==0.14.5
tenacity==8.2.3
